fix: Address review feedback on ReadBlock component
- Use highlightBlock on full content then split by newline to preserve
  multi-line token context (block comments, template literals)
- Skip syntax highlighting for error state, render as plain escaped text
- Extract EXT_TO_LANG and getLanguage to highlighting.js to deduplicate
  with DiffView.svelte
- Fix chevron to use consistent {expanded ? '▾' : '▸'} like siblings
